Rename fasta source to fasta providers
